pyproject.toml
[project]
name = "analytics-mcp"
version = "0.1.0"
description = "Combined Google Analytics and Search Console MCP server"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "mcp[cli]>=1.3.0",
    "google-analytics-data>=0.20.0,<1.0",
    "google-analytics-admin>=0.27.0,<1.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.20.0",
    "google-auth-httplib2>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4.0",
]

[project.scripts]
analytics-mcp = "analytics_mcp.server:run_server"

[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"
